Fix premature closure of TCP connection
Fixes: #152 Premature closure of TCP connection

Clear the FIN flag for packets that are being dropped during so_rcv buffer
full condition.

This patch fixes a problem introduced by commit
af101aa4e648b020886fc36b51df4cc80e2c98a9 for bug #93 (Data corruption - tcp
packets dropped in the receive path).
If there is not enough room in the so_rcv sockbuf, the fix of bug #93
correctly drops the current packet to trigger a retransmit.
The problem occurs if that packet happens to be a FIN, in which case the
packet gets dropped but its FIN flag still gets processed, causing the
connection to close prematurely.
The patch below fixes the problem by clearing the FIN flag for such packets.

Signed-off-by: Janos Villanyi <a74esev@gmail.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>