Removal of key from ARP CACHE to make arp cache update atomic
This is a fix for Git Issue 121: Race condition in MAC address lookup.
Writing and reading a ARP cache entry was not atomic.
By removing the key from the structure we can make writing
and reading of cache atomic. As each cache points to ARP entry so to
check whether its a hit or not we compare the key with key present
in the entry.

Signed-off-by: Sirshak Das <sirshak.das@arm.com>
Reviewed-by: Brian Brooks <brian.brooks@arm.com>
Reviewed-by: Honnappa Nagarahalli <honnappa.nagarahalli@arm.com>
Reviewed-by: Sorin Vultureanu <sorin.vultureanu@enea.com>